refactor: apply hexagonal architecture and openrewrite tooling
