+++ rewrite.yml
@@ -1,0 +1,4 @@
+type: specs.openrewrite.org/v1beta/recipe
+name: com.example.hexagonal.RewriteRecipes
+recipeList:
+  - org.openrewrite.java.format.AutoFormat
