+++ src/main/java/org/shark/renovatio/McpServerApplication.java
@@ -9,23 +9,3 @@ public class McpServerApplication {
         SpringApplication.run(McpServerApplication.class, args);
     }
 }
-package org.shark.renovatio.model;
-
-public class RefactorRequest {
-    private String sourceCode;
-    private String recipe;
-
-    public String getSourceCode() {
-        return sourceCode;
-    }
-    public void setSourceCode(String sourceCode) {
-        this.sourceCode = sourceCode;
-    }
-    public String getRecipe() {
-        return recipe;
-    }
-    public void setRecipe(String recipe) {
-        this.recipe = recipe;
-    }
-}
-
